+++ ingest-tools/fmp4stream.h
@@ -300,7 +300,7 @@ namespace fMP4Stream {
 		
 		void write_emsg_as_fmp4_fragment(std::ostream &out, uint64_t tfdt, uint32_t track_id, uint64_t next_tfdt, uint8_t target_version); // warning may change the version
 		void write_emsg_as_mpd_event(std::ostream &ostr, uint64_t base_time) const;
-		void convert_emsg_to_sparse_fragment(std::vector<uint8_t> &sparse_frag_out, uint32_t tfdt, uint32_t track_id, uint32_t timescale, uint8_t target_emsg_version = 0);
+		void convert_emsg_to_sparse_fragment(std::vector<uint8_t> &sparse_frag_out, uint64_t tfdt, uint32_t track_id, uint32_t timescale, uint8_t target_emsg_version = 0);
 	};	
 
 	const uint8_t empty_mfra[8] = {
